+++ src/fw_iptables.c
@@ -916,11 +916,13 @@ iptables_download_ratelimit_enable(t_client *client, int enable)
 		bucket = 5;
 	} else {
 		average_packet_size = (client->counters.incoming - client->counters.incoming_previous) /
-			(client->counters.inpackets - client->counters.inpackets_previous);
+			(client->counters.inpackets - client->counters.inpackets_previous) + 1;
 		packet_limit = client->download_rate * 1024 * 60 / average_packet_size / 8; // packets per minute
 		packets = client->downrate * 1024 * 60 / average_packet_size / 8; // packets per minute
 		bucket = (packets - packet_limit) / packet_limit;
 	}
+
+	bucket = bucket * config->download_bucket_ratio;
 
 	if ( bucket < 5) {
 		bucket = 5;
@@ -930,8 +932,6 @@ iptables_download_ratelimit_enable(t_client *client, int enable)
 	if ( packet_limit > 300000) {
 		packet_limit = 300000;
 	}
-
-	bucket = bucket * config->download_bucket_ratio;
 
 	if ( bucket > config->max_download_bucket_size) {
 		bucket = config->max_download_bucket_size;
@@ -1003,11 +1003,13 @@ iptables_upload_ratelimit_enable(t_client *client, int enable)
 		bucket = 5;
 	} else {
 		average_packet_size = (client->counters.outgoing - client->counters.outgoing_previous) /
-			(client->counters.outpackets - client->counters.outpackets_previous);
+			(client->counters.outpackets - client->counters.outpackets_previous) + 1;
 		packet_limit = client->upload_rate * 1024 * 60 / average_packet_size / 8; // packets per minute
 		packets = client->uprate * 1024 * 60 / average_packet_size / 8; // packets per minute
 		bucket = (packets - packet_limit) / packet_limit;
 	}
+
+	bucket = bucket * config->upload_bucket_ratio;
 
 	if ( bucket < 5) {
 		bucket = 5;
@@ -1017,8 +1019,6 @@ iptables_upload_ratelimit_enable(t_client *client, int enable)
 	if ( packet_limit > 300000) {
 		packet_limit = 300000;
 	}
-
-	bucket = bucket * config->upload_bucket_ratio;
 
 	if ( bucket > config->max_upload_bucket_size) {
 		bucket = config->max_upload_bucket_size;